+++ src/pyownet/__init__.py
@@ -18,7 +18,7 @@
 #
 
 __all__ = ['__version__', 'Error']
-__version__ = '0.9.1.dev1'
+__version__ = '0.9.1.dev2'
 
 
 class Error(Exception):